Added upgrades for Solar Park, and finished upgrades for Coal Plant. Made Solar Parks daytime-dependent

+++ Assets/Scripts/Core Logic/SaveState.cs
@@ -6,7 +6,7 @@ public class SaveState
 {
     public long money;
     public float happiness;
-    public int emissions;
+    public float emissions;
 
     public List<TileSaveData> tiles;
 }
